support specifying lpic in DESIGNATE_RPIC
Add support for passing label name as second parameter to
DESIGNATE_RPIC:
If second parameter to DESIGNATE_RPIC is not a number, it's treated
as label name and lpic is [re-]defined as (%2) and lpiccf flag is set to
1.

Add check for cases where lpiccache address uses rpic register (as base
or index), during cache update.

Loading from cache in the same situation is OK, e.g.:
    ...           ; somehow write $$ to [ebp+4]
    %xdefine lpiccache [ebp+4]
    %xdefine lpic      $$
    %xdefine lpiccf    1
    PIC_BEGIN ebp ; load from cache: mov ebp, [ebp+4]

Add rpic init/load comments.